client: Save password with unique identificator
Since the cient did not save the users credentials under a unique
identificator in the keyring, credentials got mixed up. This lead
to situations where the user was not prompted for a password, but
authentication failed. Added attribute keyring_identificator to
the ProfitBricksService class to be used when reading/saving
passwords to the keyring.